feat(but): replace old commit command with commit2
