+++ crates/but/src/args/commit.rs
@@ -1,118 +1,97 @@
+//! Arguments for `commit`.
+
+#![deny(missing_docs)]
+
 use crate::args::atoms::CliIdArg;
 
+/// Create a commit.
+///
+/// By default, all uncommitted changes are included in the commit. This can be controlled with
+/// change flags such as `--empty` and `--interactive`, or by providing `CHANGES` as positional
+/// arguments.
+///
+/// If there are no branches applied, a new branch is created for the commit. If there is only one
+/// branch applied, the commit is placed at the tip of that branch. Otherwise, the targeting flags
+/// `--above`, `--below` and `--branch` must be used to control where the commit is placed. Note
+/// that only one of the targeting flags can be provided at a time.
+///
+/// The commit is expected to have a commit message unless `--no-message` is provided. If neither of
+/// `--no-message` nor `--message` is provided, the user's preferred editor is opened to input a
+/// message.
 #[derive(Debug, clap::Parser)]
+#[cfg_attr(feature = "raw-clap-docs", clap(verbatim_doc_comment))]
 pub struct Platform {
-    /// Commit message
-    #[clap(short = 'm', long = "message", conflicts_with = "message_file")]
-    pub message: Option<String>,
-    /// Read commit message from file
-    #[clap(long = "message-file", value_name = "FILE", conflicts_with = "message")]
-    pub message_file: Option<std::path::PathBuf>,
-    /// Branch CLI ID or name to derive the stack to commit to
-    pub branch: Option<CliIdArg>,
-    /// Whether to create a new branch for this commit.
-    /// If the branch name given matches an existing branch, that branch will be used instead.
-    /// If no branch name is given, a new branch with a generated name will be created.
-    #[clap(short = 'c', long = "create")]
-    pub create: bool,
-    /// Insert the commit before this commit or branch.
-    #[clap(long, conflicts_with = "after")]
-    pub before: Option<CliIdArg>,
-    /// Insert the commit after this commit or branch.
-    #[clap(long, conflicts_with = "before")]
-    pub after: Option<CliIdArg>,
-    /// No-op compatibility flag for `git commit -a`.
-    #[clap(short = 'a', long = "all")]
-    pub all: bool,
-    /// Bypass pre-commit hooks
-    #[clap(short = 'n', long = "no-hooks", alias = "no-verify")]
-    pub no_hooks: bool,
-    /// Generate commit message using AI with optional user summary.
-    /// Use --ai by itself or --ai="your instructions" (equals sign required for value)
-    #[clap(
-        short = 'i',
-        long = "ai",
-        conflicts_with = "message",
-        conflicts_with = "message_file",
-        num_args = 0..=1,
-        require_equals = true
-    )]
-    pub ai: Option<Option<String>>,
-    /// Uncommitted file or hunk CLI IDs to include in the commit.
-    /// Can be specified multiple times or as comma-separated values.
-    /// If not specified, all uncommitted changes (or changes staged to the target branch) are committed.
-    #[clap(long = "changes", short = 'p', value_delimiter = ',')]
-    pub changes: Vec<String>,
-    /// Always show diff inside the editor.
+    /// The message to use for the commit.
     ///
-    /// By default the diff will be shown unless it's large. The diff will always be shown if
-    /// `--diff` is passed, regardless of the size of the diff.
-    #[clap(long = "diff", default_value_t, conflicts_with_all = &["no_diff", "message", "message_file", "ai"])]
-    pub diff: bool,
-    /// Never show the diff inside the editor.
-    #[clap(long = "no-diff", default_value_t, conflicts_with_all = &["diff", "message", "message_file", "ai"])]
-    pub no_diff: bool,
-    #[clap(subcommand)]
-    pub cmd: Option<Subcommands>,
-}
+    /// Can be supplied any amount of times, each value being appended to the preceding ones with a
+    /// blank line in between.
+    #[clap(short, long, group = "commit_message")]
+    pub message: Option<Vec<String>>,
 
-#[derive(Debug, clap::Subcommand)]
-pub enum Subcommands {
-    /// Insert a blank commit before or after the specified commit.
+    /// Creates the commit without a commit message.
+    #[clap(long, group = "commit_message")]
+    pub no_message: bool,
+
+    /// Place the commit on the branch `BRANCH`.
     ///
-    /// This is useful for creating a placeholder commit that you can
-    /// then amend changes into later using `but rub` or `but absorb`.
+    /// If `BRANCH` does not exist, it is created as an unstacked branch.
     ///
-    /// You can provide a message with `-m` or modify it later using `but reword`.
+    /// If `BRANCH` is omitted, an unstacked branch with a generated name is created.
     ///
-    /// This allows for a more Jujutsu style workflow where you create commits
-    /// first and then fill them in as you work. Create an empty commit and
-    /// then rub or absorb changes into it whenever you prefer.
+    /// Attempting to place a commit on a branch that exists but is not applied is an error.
     ///
-    /// ## Examples
+    /// Takes a CLI ID, see `but help cli-ids` for details.
+    #[clap(short, long, value_name = "BRANCH", group = "targeting")]
+    pub branch: Option<Option<CliIdArg>>,
+
+    /// Place the commit above `BRANCH_OR_COMMIT`, which must be an applied branch or commit.
     ///
-    /// Insert at the top of the first branch (no arguments):
+    /// If `BRANCH_OR_COMMIT` is a commit, the new commit is placed on the same branch as the
+    /// targeted commit.
     ///
-    /// ```text
-    /// but commit empty
-    /// ```
+    /// If `BRANCH_OR_COMMIT` is a branch, the new commit is placed on a new branch above the
+    /// targeted branch.
     ///
-    /// Insert before a commit:
+    /// Takes a CLI ID, see `but help cli-ids` for details.
+    #[clap(
+        short = 'A',
+        long,
+        value_name = "BRANCH_OR_COMMIT",
+        group = "targeting"
+    )]
+    pub above: Option<CliIdArg>,
+
+    /// Place the commit below `BRANCH_OR_COMMIT`, which must be an applied branch or commit.
     ///
-    /// ```text
-    /// but commit empty ab
-    /// ```
+    /// If `BRANCH_OR_COMMIT` is a commit, the new commit is placed on the same branch as the targeted
+    /// commit.
     ///
-    /// Explicitly insert before a commit:
+    /// If `BRANCH_OR_COMMIT` is a branch, the new commit is placed on a new branch below the
+    /// targeted branch. Branches are treated as buckets, meaning that "below a branch" is treated
+    /// as below the oldest ancestor on that branch.
     ///
-    /// ```text
-    /// but commit empty --before ab
-    /// ```
+    /// Takes a CLI ID, see `but help cli-ids` for details.
+    #[clap(
+        short = 'B',
+        long,
+        value_name = "BRANCH_OR_COMMIT",
+        group = "targeting"
+    )]
+    pub below: Option<CliIdArg>,
+
+    /// Forces the commit to be empty regardless of repository state.
+    #[clap(long, group = "changes_to_commit")]
+    pub empty: bool,
+
+    /// Open the TUI to interactively select what to commit.
+    #[clap(short, long, group = "changes_to_commit")]
+    pub interactive: bool,
+
+    /// One or more changes to commit.
     ///
-    /// Insert after a commit (at the top of the stack if target is a branch):
+    /// A change can either be a file or a hunk.
     ///
-    /// ```text
-    /// but commit empty --after ab
-    /// ```
-    ///
-    #[cfg(feature = "legacy")]
-    #[cfg_attr(feature = "raw-clap-docs", clap(verbatim_doc_comment))]
-    #[command(group = clap::ArgGroup::new("position"))]
-    Empty {
-        /// The target commit or branch to insert relative to.
-        ///
-        /// If a target is provided without --before or --after, defaults to --before behavior.
-        /// If no arguments are provided at all, inserts at the top of the first branch.
-        #[arg(group = "position")]
-        target: Option<CliIdArg>,
-        /// Insert the blank commit before this commit or branch
-        #[arg(long, group = "position")]
-        before: Option<CliIdArg>,
-        /// Insert the blank commit after this commit or branch
-        #[arg(long, group = "position")]
-        after: Option<CliIdArg>,
-        /// Commit message for the inserted blank commit
-        #[arg(short = 'm', long = "message")]
-        message: Option<String>,
-    },
+    /// Takes CLI IDs, see `but help cli-ids` for details.
+    #[clap(group = "changes_to_commit")]
+    pub changes: Vec<CliIdArg>,
 }
