+++ crates/but/src/lib.rs
@@ -455,10 +455,7 @@ async fn match_subcommand(
             if let Ok(props) = utils::metrics::Props::from_json_string(&props) {
                 props.update_event(&mut event);
             }
-            if matches!(
-                command_name,
-                CommandName::Commit | CommandName::CommitEmpty | CommandName::Commit2
-            ) {
+            if matches!(command_name, CommandName::Commit | CommandName::CommitEmpty) {
                 utils::metrics::add_workspace_shape(&mut event, &args.current_dir);
             }
             utils::metrics::capture_event_blocking(&app_settings, event).await;
@@ -1025,139 +1022,6 @@ async fn match_subcommand(
         }
         #[cfg(feature = "legacy")]
         Subcommands::Commit(commit_args) => {
-            if let Ok(repo) = gix::discover(&args.current_dir)
-                && repo.worktree().is_some_and(|worktree| !worktree.is_main())
-            {
-                return Err(anyhow::anyhow!(
-                    "`but commit` is not supported from linked worktrees. Use Git directly for this worktree."
-                )
-                .into());
-            }
-
-            let status_after = args.status_after;
-            let mut ctx = setup::init_ctx(
-                &args,
-                InitCtxOptions {
-                    background_sync: BackgroundSync::Enabled { silent: false },
-                    ..Default::default()
-                },
-                out,
-            )?;
-            out.begin_status_after(status_after);
-
-            let result = match &commit_args.cmd {
-                Some(crate::args::commit::Subcommands::Empty {
-                    target,
-                    before,
-                    after,
-                    message,
-                }) => {
-                    // Validate that no regular commit options are specified with the empty subcommand
-                    if commit_args.message.is_some() {
-                        return Err(bad_input(
-                            "--message must be passed after 'empty'. Use `but commit empty -m \"message\"`."
-                        ).into());
-                    }
-                    if commit_args.message_file.is_some() {
-                        return Err(bad_input(
-                            "--message-file cannot be used with 'commit empty'. Use `but commit empty -m \"message\"`."
-                        ).into());
-                    }
-                    if commit_args.branch.is_some() {
-                        return Err(bad_input(
-                            "branch argument cannot be used with 'commit empty'. Use the target positional argument or --before/--after flags."
-                        ).into());
-                    }
-                    if commit_args.create {
-                        return Err(
-                            bad_input("--create cannot be used with 'commit empty'.").into()
-                        );
-                    }
-                    if commit_args.before.is_some() || commit_args.after.is_some() {
-                        return Err(bad_input(
-                            "--before/--after must be passed after 'empty'. Use `but commit empty --before <target>`."
-                        ).into());
-                    }
-                    if commit_args.all {
-                        return Err(bad_input("--all cannot be used with 'commit empty'.").into());
-                    }
-                    if commit_args.no_hooks {
-                        return Err(
-                            bad_input("--no-hooks cannot be used with 'commit empty'.").into()
-                        );
-                    }
-                    if commit_args.ai.is_some() {
-                        return Err(bad_input("--ai cannot be used with 'commit empty'.").into());
-                    }
-                    if commit_args.diff {
-                        return Err(bad_input("--diff cannot be used with 'commit empty'.").into());
-                    }
-                    if commit_args.no_diff {
-                        return Err(
-                            bad_input("--no-diff cannot be used with 'commit empty'.").into()
-                        );
-                    }
-                    // Note: --paths with commit empty is rejected by clap at parse time
-                    // because --paths is not a flag on the empty subcommand
-
-                    command::legacy::commit::insert_blank_commit(
-                        &mut ctx,
-                        out,
-                        target.clone(),
-                        before.clone(),
-                        after.clone(),
-                        message.as_deref(),
-                    )
-                    .emit_metrics(metrics_ctx)
-                }
-                None => {
-                    // Handle the regular `but commit` command
-
-                    // Formats without an interactive editor must provide the message up front
-                    if !args.format.format.allows_human_ui()
-                        && commit_args.message.is_none()
-                        && commit_args.message_file.is_none()
-                        && commit_args.ai.is_none()
-                    {
-                        return Err(bad_input(
-                            "Either --message (-m), --message-file, or --ai (-i) must be specified for this output format"
-                        ).into());
-                    }
-
-                    // Read message from file if provided, otherwise use message option
-                    let commit_message = match &commit_args.message_file {
-                        Some(path) => Some(std::fs::read_to_string(path).with_context(|| {
-                            format!(
-                                "Failed to read commit message from file: {}",
-                                path.display()
-                            )
-                        })?),
-                        None => commit_args.message.clone(),
-                    };
-                    command::legacy::commit::commit(
-                        &mut ctx,
-                        out,
-                        commit_message.as_deref(),
-                        commit_args.branch.clone(),
-                        commit_args.before.clone(),
-                        commit_args.after.clone(),
-                        &commit_args.changes,
-                        commit_args.all,
-                        commit_args.create,
-                        commit_args.no_hooks,
-                        commit_args.ai.clone(),
-                        ShowDiffInEditor::from_args(commit_args.diff, commit_args.no_diff)
-                            .unwrap_or(ShowDiffInEditor::Unspecified),
-                    )
-                    .emit_metrics(metrics_ctx)
-                }
-            };
-
-            run_status_after_if_ok(status_after, &result, &mut ctx, out);
-            result
-        }
-        #[cfg(feature = "legacy")]
-        Subcommands::_Commit2(commit_args) => {
             use crate::utils::IntermediateChannel;
 
             let status_after = args.status_after;
@@ -1171,7 +1035,7 @@ async fn match_subcommand(
             )?;
             out.begin_status_after(status_after);
 
-            let outcome = command::legacy::commit2::commit(
+            let outcome = command::legacy::commit::commit(
                 &mut ctx,
                 IntermediateChannel::new(out),
                 commit_args,
