+++ crates/but/tests/but/command/reword.rs
@@ -46,7 +46,7 @@ fn reword_commit_with_change_id_shows_change_id() {
     env.setup_metadata(&["A"]);
 
     env.file("new.txt", "content\n");
-    env.but("commit A -m 'add new.txt'").assert().success();
+    env.but("commit -b A -m 'add new.txt'").assert().success();
 
     env.but("reword 1 -m 'reworded'")
         .assert()
